django/requirements.txt
Django>=5.2.0,<6.0
djangorestframework>=3.16.0,<4.0
psycopg2-binary>=2.9.10,<3.0
django-mptt>=0.17.0,<1.0
Pillow>=11.2.0,<12.0
openpyxl>=3.1.5,<3.2.0
gunicorn>=23.0.0,<24.0
django-model-utils>=5.0.0,<6.0
django-cors-headers>=4.3.0,<5.0
